Migrate gatsby-node to be typestrict, fix build
