+++ gatsby-config.js
@@ -35,6 +35,7 @@ module.exports = {
 			}
 		},
 		`gatsby-transformer-json`,
+		`gatsby-plugin-typescript`,
 		{
 			resolve: `gatsby-source-filesystem`,
 			options: {
